V  = 3.4.1.U
- funcionalidad para obtener los tags y las enfermedades a partir de los tags
- se elimino componente input-tags para renderizar los tags de enfermedades directamente en el formulario de consulta
