+++ app/(pages)/doctor/consultas/[consulta]/components/form-diagnostic.tsx
@@ -13,7 +13,8 @@ import { useForm } from 'react-hook-form'
 import * as z from 'zod'
 import { ToastContainer, toast } from 'react-toastify'
 import { useTransition } from 'react'
-import InputTags from './input-tags'
+import { XMarkIcon } from '@heroicons/react/24/outline'
+import { createDiagnostico } from '../../../actions'
 
 const validationSchema = z.object({
 /*   enfermedades: z.string().min(1, { message: 'Las enfermedades son requeridas' }), */
@@ -38,6 +39,59 @@ type ValidationSchema = z.infer<typeof validationSchema>
 
 export default function FormDiagnostic ({ consulta }: { consulta: Consultas }) {
   const [isPending, startTransition] = useTransition()
+  const [tags, setTags] = React.useState<string[]>([]) // useState para el arreglo de tags
+  const [tagsEnfermedades, setTagsEnfermedades] = React.useState<string>('') // useState para las enfermedades (string extraido del arreglo de tags)
+  let newTags: string[] = [] // variable auxiliar para el arreglo de tags
+  // agregar tags
+  const handleAddTag = (tag: string) => {
+    if (tag !== '') {
+      newTags = [...tags, tag]
+      setTags(newTags)
+    }
+    return newTags
+  }
+  // eliminando los tags
+  const handleRemoveTag = (index: number) => {
+    if (index !== -1) {
+      newTags = [...tags]
+      newTags.splice(index, 1)
+      setTags(newTags)
+    }
+    return newTags
+  }
+  // funcion para se crear los tags al presionar una coma
+  const handleKeyDown = (event: React.KeyboardEvent<HTMLInputElement>) => {
+    if (event.key === ',') {
+      const tag = event.currentTarget.value.trim().replace(/\s*,\s*/g, '')
+      if (tag) {
+        handleAddTag(tag)
+        event.currentTarget.value = ''
+      }
+      event.preventDefault() // Prevent the comma from being entered in the input field
+    }
+  }
+  // handle cuando se pierde el focus del input para agregar enfermedades
+  const handleBlur = (event: React.FocusEvent<HTMLInputElement>) => {
+    const tag = event.currentTarget.value.trim()
+    if (tag) {
+      handleAddTag(tag)
+      event.currentTarget.value = ''
+    }
+    const tagsSended = handleAddTag('')
+    getTags(tagsSended ? tagsSended.join(',') : '')
+  }
+  // funcion que llama a removeTags presionando x
+  const handleTagClick = (index: number) => {
+    handleRemoveTag(index)
+    const tagsSended = handleRemoveTag(-1)
+    getTags(tagsSended ? tagsSended.join(',') : '')
+  }
+  // obtenemos los tags y convertimos el arreglo a string
+  function getTags (enfermedades: string) {
+    const enfermedadesTags = enfermedades
+    console.log(enfermedadesTags)
+    setTagsEnfermedades(enfermedadesTags)
+  }
 
   const {
     register,
@@ -68,7 +122,10 @@ export default function FormDiagnostic ({ consulta }: { consulta: Consultas }) {
       // Generar fecha del diagnostico
       data.fecha_diagnostico = new Date().toISOString()
 
-      /*
+      // Añadir enfermedades
+      data.enfermedades = tagsEnfermedades
+      console.log(tagsEnfermedades)
+
       const { diagnostico, errorDiagnostico } = await createDiagnostico({
         data
       })
@@ -84,7 +141,6 @@ export default function FormDiagnostic ({ consulta }: { consulta: Consultas }) {
       if (!diagnostico) {
         toast.error('Error al crear el diagnostico')
       }
-      */
     })
   }
 
@@ -92,13 +148,7 @@ export default function FormDiagnostic ({ consulta }: { consulta: Consultas }) {
     <div className="grid gap-6">
       <form onSubmit={handleSubmit(onSubmit)}>
         <div className="grid gap-3">
-            <div className="grid gap-1">
-              <Label className="" htmlFor="enfermedades">
-                Enfermedades
-              </Label>
-              <p className="mt-1 leading-6 px-4 text-gray-600 dark:text-white">
-                Ingrese una enfermedad o varias enfermedades separadas por comas
-              </p>
+            {/* <div className="grid gap-1">
               <Input
                 type="text"
                 autoComplete="enfermedades"
@@ -116,8 +166,45 @@ export default function FormDiagnostic ({ consulta }: { consulta: Consultas }) {
                   {errors.enfermedades?.message}
                 </p>
               )}
-            </div>
-            <InputTags/>
+            </div> */}
+          <div className="grid gap-1">
+            <Label className="" htmlFor="enfermedades">
+              Enfermedades
+            </Label>
+            <p className="mt-1 leading-6 px-4 text-gray-600 dark:text-white">
+              Ingrese una enfermedad o varias enfermedades separadas por comas
+            </p>
+          </div>
+          <div className="flex gap-1  border border-solid rounded-md items-center px-2">
+            <ul
+              className=' flex gap-1'
+            >
+              { tags.map((tag, index) => (
+                <li key={index} className='text-white flex  group  bg-secundario rounded-lg  border-1 border-blue-400  px-2 py-1'>
+                  <p className=' capitalize text-sm truncate '>
+                    {tag}
+                  </p>
+                  <span
+                    className='inline-flex justify-center items-center ml-2 rounded-full transition-colors hover:bg-secundariovariant-600 p-1 cursor-pointer'
+                    onClick={() => { handleTagClick(index) }}
+                  >
+                    <XMarkIcon className="h-3 w-3" />
+                  </span>
+                </li>
+
+              ))}
+            </ul>
+            <Input
+              type="text"
+              autoComplete="enfermedades"
+              placeholder="enfermedades"
+              className='border-0 outline-none focus:!ring-0 focus:!border-0 focus:ring-offset-0  focus-visible:ring-offset-0 '
+              onKeyDown={handleKeyDown}
+              onBlur={handleBlur}
+              disabled={isPending}
+            />
+
+          </div>
             <div className="grid gap-1">
               <Label className="" htmlFor="observaciones">
                 Observaciones
